requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
stripe
joblib
scikit-learn
google-generativeai